11-9 update 10:18 by Phúc

+++ src/main/java/com/project/sem4/model/view/OrderView.java
@@ -9,12 +9,11 @@ public class OrderView {
     private Integer id;
     private String  userName;
     private Integer total;
-    private Integer totalPrice;
+    private Double totalPrice;
     private String paymentName;
     private String phone;
     private String address;
     private String description;
     private Date createDate;
     private Integer status;
-
 }
